requirements.txt
pyside6==6.6
python-i18n==0.3.9
vpython==7.6.4
pytest==7.4.3
